Build --static on windows
This small changes ensure that when building under mingw (with conda or a plain mingw) the resulting _graphillion.pyd library does not have extra dependencies that are MinGW-specific such as libstdc++ or libgcc. This allows to build further wheels and eggs that run out of the box. The (minor) downside is that the shared library is about twice bigger (1MB vs 500KB).